requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
ultralytics==8.3.7
paddleocr==2.7.3
opencv-python-headless==4.10.0.84
numpy==1.26.4
python-multipart==0.0.9
requests==2.32.3
paddlepaddle==2.6.1
python-dotenv==1.1.0
protobuf==3.20.2
mangum==0.17.0